requirements.txt
awscli==1.14.49
Babel==2.5.0
boto==2.48.0
boto3==1.6.2
botocore==1.9.2
backports.shutil-get-terminal-size==1.0.0
cloudpickle==0.5.2
colorama==0.3.7
colorlover==0.2.1
Cython==0.26.1
dash==0.19.0
dash-core-components==0.15.0rc1
dash-html-components==0.8.0
dash-renderer==0.11.1
dash-table-experiments==0.5.0
dask==0.16.0
docutils==0.14
dropbox==8.5.0
fastcache==1.0.2
Flask-Cache==0.13.1
Flask-Caching==1.3.3
Flask==0.12.2
fastcache==1.0.2
Flask-Compress==1.4.0
Flask-Cors==3.0.3
fuzzywuzzy==0.15.1
gensim==3.0.1
gevent==1.2.2
gunicorn==19.7.1
google-api-python-client==1.6.4
googlemaps==2.5.1
h5py==2.7.1
heapdict==1.0.0
html5lib==0.999999999
httplib2==0.10.3
hyperlink==17.1.1
hypothesis==3.37.0
imageio==2.2.0
imagesize==0.7.1
incremental==17.5.0
isort==4.2.15
inflect==0.2.5
intrinio==0.1.17
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.10.2
jellyfish==0.3.2
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
language-check==1.1
matplotlib==2.1.0
nltk==3.2.5
numpy==1.13.3
numpydoc==0.7.0
oauth2client==4.1.2
onedrivesdk==1.1.8
openpyxl==2.4.8
pandas==0.21.0
pandas-datareader==0.5.0
pandocfilters==1.4.2
pickleshare==0.7.4
Pillow==4.2.1
plotly==2.2.1
psycopg2==2.7.3.1
ptyprocess==0.5.2
py==1.5.2
PyDrive==1.3.1
pyflakes==1.6.0
Pygments==2.2.0
pyLDAvis==2.1.1
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.2.5
pyflakes==1.6.0
pyOpenSSL==17.2.0
python-dateutil==2.6.1
python-google-places==1.4.1
python-Levenshtein==0.12.0
python-onedrive==14.9.3
queuelib==1.4.2
redis==2.10.6
regex==2017.11.9
requests==2.18.4
requests-file==1.4.2
requests-ftp==0.3.1
rope==0.10.7
scikit-image==0.13.1
scikit-learn==0.19.1
scikit-plot==0.3.1
s3transfer==0.1.13
scipy==1.0.0
Scrapy==1.4.0
seaborn==0.8.1
selenium==3.7.0
service-identity==17.0.0
spacy==1.9.0
Shapely==1.6.4.post1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
squarify==0.3.0
statsmodels==0.8.0
six==1.11.0
sumy==0.7.0
sympy==1.1.1
tables==3.4.2
tqdm==4.19.4
ujson==1.35
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
w3lib==1.18.0
xgboost==0.6a2
xlrd==1.1.0
XlsxWriter==1.0.2